week3/rag_streamlit_pdf/requirements.txt
streamlit>=1.37
faiss-cpu>=1.8.0
sentence-transformers>=3.0.1
pypdf>=4.2.0
langchain-text-splitters>=0.2.2
openai>=1.35.0
python-dotenv>=1.0.1
tiktoken>=0.7.0
